Further Progress On Item Checkout
Fixed the database to convert all spaces into dashes, and convert
all dashes into spaces when being displayed in the template via
a template tag.

Adjusted item removal URL to display the item in question that
is being 'checked out'.

Going in the direction of having items only to be checked out
via actually searching for them. Doesn't make sense to have a page
which allows a user to input an item to be checked out instead of
doing so from the search page itself. Think Amazon

Need to refactor instances of 'part' into just 'item'